fix: ErrorCode 변경에 따른 빌드 실패 해결
