+++ src/main/java/com/airbnb_clone/chatting/controller/ChatRoomController.java
@@ -1,12 +1,5 @@
 package com.airbnb_clone.chatting.controller;
 
-import com.airbnb_clone.chatting.common.global.response.ApiResponse;
-import com.airbnb_clone.chatting.repository.Dto.chatRoom.ChatRoomNewReqDto;
-import com.airbnb_clone.chatting.repository.Dto.chatRoom.ChatRoomNewResDto;
-import com.airbnb_clone.chatting.service.ChatRoomService;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.bson.types.ObjectId;
 import org.springframework.http.HttpEntity;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -14,6 +7,14 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+
+import com.airbnb_clone.chatting.repository.Dto.chatRoom.ChatRoomNewReqDto;
+import com.airbnb_clone.chatting.repository.Dto.chatRoom.ChatRoomNewResDto;
+import com.airbnb_clone.chatting.service.ChatRoomService;
+import com.airbnb_clone.common.global.response.ApiResponse;
+
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 
 @Slf4j
 @RestController
@@ -28,7 +29,7 @@ public class ChatRoomController {
         ChatRoomNewResDto save = chatRoomService.save(chatRoomNewReqDto);
 
         return ResponseEntity.ok(
-                ApiResponse.of("생성 성공!", HttpStatus.OK.value(), save)
+                ApiResponse.of("생성 성공!", HttpStatus.OK, save)
         );
     }
