+++ src/main/java/com/airbnb_clone/exception/ErrorResponse.java
@@ -1,6 +1,7 @@
 package com.airbnb_clone.exception;
 
 import lombok.AccessLevel;
+import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
@@ -9,35 +10,17 @@ import lombok.NoArgsConstructor;
 public class ErrorResponse {
     private String message;
     private int status;
-
+    
+    @Builder
     private ErrorResponse(String message, int status) {
         this.message = message;
         this.status = status;
     }
 
     public static ErrorResponse of(ErrorCode errorCode) {
-        return new Builder()
-                .message(errorCode.getMessage())
-                .status(errorCode.getStatus())
-                .build();
-    }
-
-    static class Builder {
-        private String message;
-        private int status;
-
-        public Builder message(String message) {
-            this.message = message;
-            return this;
-        }
-
-        public Builder status(int status) {
-            this.status = status;
-            return this;
-        }
-
-        public ErrorResponse build() {
-            return new ErrorResponse(message, status);
-        }
+        return ErrorResponse.builder()
+            .message(errorCode.getMessage())
+            .status(errorCode.getStatus().value())
+            .build();
     }
 }
